functions/requirements copy.txt
google-cloud-aiplatform>=1.36.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0 
google-adk==1.12.0
firecrawl-py==4.3.6
flask==2.3.2
fastapi==0.116.1
uvicorn==0.34.0
supabase>=2.0.0
python-multipart>=0.0.9
litellm==1.77.4
googlemaps==4.10.0
yt-dlp==2025.10.14
google-genai==1.45.0
google-generativeai==0.8.5
google-auth==2.41.1
google-cloud-core==2.4.3
alembic==1.17.1
httpx==0.28.1
serpapi==0.1.5
anchorbrowser==0.3.12